requirements.txt
openai 
phidata
python-dotenv
yfinance
duckduckgo-search
fastapi 
uvicorn
groq
